multi-ai-platform-manager/requirements.txt
# 基础依赖
playwright>=1.40.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio>=3.4.3

# 可选OCR引擎
pytesseract  # 需要安装系统Tesseract

# 开发工具
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 配置管理
python-dotenv>=1.0.0